+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). The date format used is DD/MM/YYYY.
 
+## [3.3.0] - 01/11/2025
+### Added
+* New certificate: UX Management: Strategy and Tactics
+* New certificate: How to Design Engaging Products: Insights from Fortnite's UX
+
 ## [3.2.9] - 08/08/2025
 ### Added
 * New certificate: How to Elevate the User Experience of AI with Design Patterns